440: Split ballot paper checks into separate files
