+++ apps/backend/src/controllers/ballotPapers/updateBallotPaper.uc.ts
@@ -8,7 +8,7 @@ import {
 import type { Request, Response } from 'express';
 import { HttpStatusCode } from '../../httpStatusCode.js';
 import { updateBallotPaper as updatePersistentBallotPaper } from '../../services/ballotPapers.service.js';
-import { validateUpdateableBallotPaper } from '../.bodyChecks/ballotPaperChecks.js';
+import { validateUpdateableBallotPaper } from '../.bodyChecks/ballotPaperChecks/updatableBallotPaper.check.js';
 import { isBodyCheckValidationError } from '../.bodyChecks/bodyCheckValidationError.js';
 
 export const updateBallotPaper = async (
