+++ apps/backend/test/ballotPapers/updateBallotPaper.test.ts
@@ -10,7 +10,7 @@ import request from 'supertest';
 import { beforeAll, describe, expect, it } from 'vitest';
 import { app } from '../../src/app.js';
 import { generateUserTokens } from '../../src/auth/utils.js';
-import { BallotPaperBodyCheckValidationErrorMessage } from '../../src/controllers/.bodyChecks/ballotPaperChecks.js';
+import { BallotPaperBodyCheckValidationErrorMessage } from '../../src/controllers/.bodyChecks/ballotPaperChecks/updatableBallotPaper.check.js';
 import { HttpStatusCode } from '../../src/httpStatusCode.js';
 import { createUser, findUserBy } from '../../src/services/users.service.js';
 import {
